Back button on Android device has different behavior than function "history.back"
implemented a hashchange listener and the jshistory object now also hold on the which page/formcontroller has a hash value. And will return to that controller if the current form is not the current hash.
Try to revert to the form by reverting to the current controller if the form didn't allow a change
Case: SVY-4284