[Mobile] Sync (when there are no local changes) deletes all data if loading data fails
for now just do the clean later, after the first request for the offline datadescription. If that one works then we will clear it and get the rest. Else it could take to much memory on the mobile because the whole dataset must be in memory twice. if the first request is fine and the offline data description package is received just fine then it is more then likely that the rest will follow just fine.
Case: SVY-4438