Loading data after logout doesn't trigger login form
only removeSearch should revert to the right form when a login was shown (after a success call)

the errorHandler shouldn't be called for authentication failures this should just be handled by servoy and the login should be shown again.

but for this all to work we need to kill the line "setIncludeCredentials(true)" in OfflineDataProxy
because that results in that the browser takes over the completely credentials part if a autentication failure happens.

Case: SVY-5063