Reverted SVY-6546 commit : application.getUserProperty is not working in Mobile client
If you set a user property in the service solution you will/should be able to get it back in a next call from the same client.
And if you set a user property in client (mobile) solution you will/should be able to get it back when called from inside the client solution.
Case: SVY-6546