Fixed for SVY-6661 : Regression: Run JSUnit Tests on mobile solutions stopped working
I didn't manage to tune just timeouts and rely only on disk changes for
redeployment. So I had to:
- undeploy first to make sure we don't access the URL while the app. was
half redeployed (this seems to crash redeploy in Tomcat > 7 and leave
dir in undecided state)
- deploy and watch file system for signs that it was deployed
- verify over HTTP that war is available; just relying on HDD changes
was not enough, many times you would get a 404 in the opened browser

Now it never failed to run for me - but it takes around 20 sec to do the
whole undeploy/deploy/verify thing. It would probably be faster if we
change in the future to use the tomcat "manager" app  URLs and code
similar to the ones in tomcat deploy ant task, but didn't do that now as
then we would have to be really careful not to have that available in
app. server deployment - cause it would allow anyone access to
deploy/undeploy. If needed in the future I guess we should define the
Tomcat user that is required with the required rights and only deploy
the "manager" app. once when developer needs it and make sure it's no
longer deployed in app. server start.